pyproject.toml
[tool.poetry]
name = "test-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.2"
pytest-django = "^4.9.0"
mixer = "^7.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core_settings.base_settings"
python_files = [ "testing/test_*.py"]
